moved parsing c files to parsing folder
