+++ src/parsing/utils.c
@@ -6,12 +6,12 @@
 /*   By: tde-brui <tde-brui@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2023/08/17 17:43:42 by tde-brui      #+#    #+#                 */
-/*   Updated: 2023/08/17 17:44:22 by tde-brui      ########   odam.nl         */
+/*   Updated: 2023/08/19 15:08:39 by tde-brui      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../incl/tokenizer.h"
-#include "../libft/libft.h"
+#include "../../incl/tokenizer.h"
+#include "../../libft/libft.h"
 
 void	init_token(t_token *token)
 {
